// lib/api/schemas/stageTemplateResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { StageTemplateResponseDescription } from './stageTemplateResponseDescription';
import type { StageTemplateResponseDefaultContentItem } from './stageTemplateResponseDefaultContentItem';
import type { StageTemplateResponseDefaultConfig } from './stageTemplateResponseDefaultConfig';

export interface StageTemplateResponse {
  name: string;
  description?: StageTemplateResponseDescription;
  type: string;
  default_content: StageTemplateResponseDefaultContentItem[];
  default_config: StageTemplateResponseDefaultConfig;
  is_public?: boolean;
  id: string;
  created_at: string;
}